microservices/feedback-chatbot/requirements.txt
pyodbc==5.1.0
Flask==3.0.3
google-cloud-language==2.13.3